+++ CHANGES.md
@@ -38,6 +38,9 @@
 - Add `Arg.Conv` module to define converters. This should be used in
   new code. 
 
+- Add `Arg.{file,dir}path` string converters equiped with appropriate
+  file system completions.
+
 - Add `docv` optional parameter to `Arg.enum`.
 
 - Use `Arg.conv`'s `docv` property in the documentation of arguments